+++ device/api/umd/device/arc/blackhole_arc_message_queue.h
@@ -5,7 +5,7 @@
  */
 #pragma once
 
-#include "umd/device/blackhole_implementation.h"
+#include "umd/device/arch/blackhole_implementation.h"
 #include "umd/device/tt_core_coordinates.h"
 #include "umd/device/types/blackhole_arc.h"
 #include "umd/device/types/cluster_descriptor_types.h"
