+++ device/api/umd/device/arc/wormhole_arc_telemetry_reader.h
@@ -7,7 +7,7 @@
 
 #include "umd/device/arc/arc_messenger.h"
 #include "umd/device/arc/arc_telemetry_reader.h"
-#include "umd/device/wormhole_implementation.h"
+#include "umd/device/arch/wormhole_implementation.h"
 
 extern bool umd_use_noc1;
