+++ device/api/umd/device/arch/architecture_implementation.h
