+++ device/api/umd/device/arch/blackhole_implementation.h
@@ -9,7 +9,7 @@
 #include <array>
 #include <stdexcept>
 
-#include "umd/device/architecture_implementation.h"
+#include "umd/device/arch/architecture_implementation.h"
 #include "umd/device/types/cluster_descriptor_types.h"
 #include "umd/device/types/tlb.h"
 #include "umd/device/umd_utils.h"
