+++ device/api/umd/device/coordinates/coordinate_manager.h
