+++ device/api/umd/device/tt_soc_descriptor.h
@@ -18,7 +18,7 @@
 
 #include "fmt/core.h"
 #include "tt_xy_pair.h"
-#include "umd/device/coordinate_manager.h"
+#include "umd/device/coordinates/coordinate_manager.h"
 #include "umd/device/tt_core_coordinates.h"
 #include "umd/device/tt_xy_pair.h"
 #include "umd/device/types/arch.h"
