+++ device/arc/wormhole_arc_messenger.cpp
@@ -8,8 +8,8 @@
 #include <tt-logger/tt-logger.hpp>
 
 #include "assert.hpp"
+#include "umd/device/arch/wormhole_implementation.h"
 #include "umd/device/tt_device/tt_device.h"
-#include "umd/device/wormhole_implementation.h"
 
 extern bool umd_use_noc1;
