+++ device/chip/remote_chip.cpp
@@ -9,8 +9,8 @@
 #include <tt-logger/tt-logger.hpp>
 
 #include "assert.hpp"
+#include "umd/device/arch/wormhole_implementation.h"
 #include "umd/device/chip/local_chip.h"
-#include "umd/device/wormhole_implementation.h"
 
 namespace tt::umd {
