+++ device/coordinates/blackhole_coordinate_manager.cpp
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-#include "umd/device/blackhole_coordinate_manager.h"
+#include "umd/device/coordinates/blackhole_coordinate_manager.h"
 
 #include <tt-logger/tt-logger.hpp>
