+++ tests/api/test_chip.cpp
@@ -15,7 +15,7 @@
 #include "tests/test_utils/generate_cluster_desc.hpp"
 
 // TODO: change to tt_cluster
-#include "umd/device/architecture_implementation.h"
+#include "umd/device/arch/architecture_implementation.h"
 #include "umd/device/cluster.h"
 #include "umd/device/tt_cluster_descriptor.h"
