+++ tests/baremetal/test_core_coord_translation_wh.cpp
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 #include "gtest/gtest.h"
-#include "umd/device/coordinate_manager.h"
-#include "umd/device/wormhole_implementation.h"
+#include "umd/device/arch/wormhole_implementation.h"
+#include "umd/device/coordinates/coordinate_manager.h"
 
 using namespace tt::umd;
 #include <iostream>
