+++ tests/wormhole/test_arc_messages_wh.cpp
@@ -6,9 +6,9 @@
 
 #include "gtest/gtest.h"
 #include "umd/device/arc/arc_messenger.h"
+#include "umd/device/arch/wormhole_implementation.h"
 #include "umd/device/cluster.h"
 #include "umd/device/tt_cluster_descriptor.h"
-#include "umd/device/wormhole_implementation.h"
 
 using namespace tt::umd;
