+++ tests/wormhole/test_remote_communication_wh.cpp
@@ -4,11 +4,11 @@
 #include "gtest/gtest.h"
 #include "tests/test_utils/device_test_utils.hpp"
 #include "tests/test_utils/generate_cluster_desc.hpp"
+#include "umd/device/arch/wormhole_implementation.h"
 #include "umd/device/cluster.h"
 #include "umd/device/remote_communication.h"
 #include "umd/device/tt_cluster_descriptor.h"
 #include "umd/device/types/cluster_types.h"
-#include "umd/device/wormhole_implementation.h"
 #include "wormhole/host_mem_address_map.h"
 #include "wormhole/l1_address_map.h"
